ci/nightly: Add tests for crates main branch tracking
parsec depends on several repositories under the parallaxsecond
organization.

Currently, when testing the main branch of parsec, we are using
fixed versions of the latest released crates of those repositories.
A problem may arise when there is a change introduced in the main
branch of those repositories that would break parsec when updating/
incorporating that change. This is currently not being tested in our
CI.
As parsec release is dependent on having published the dependencies'
crates, there are issues that will not be caught until the publishing
of said crates.

* Add a nightly test to track the main branches of said repositories
  and test parsec with those updates. This test makes use of
  utils/release_tracking.py

Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>