ci/nightly-next.yml: Track and test dependencies' 'next' branch
parsec depends on several repositories under the parallaxsecond
organization.

Currently, when testing the main branch of parsec, fixed versions of
the latest released crates of those repositories are being used.
A problem may arise when there is a change introduced in the
main/'next' branch of those repositories that would break parsec when
updating/incorporating that change. This is currently not being
tested in our CI.
As parsec release is dependent on having published the dependencies'
crates, there are issues that will not be caught until the publishing
of said crates.

* Add nightly jobs that use ci.sh --next-branch-tracking to test the
  next release branches of parallaxsecond repositories and spot
  issues before release. These jobs make use of previously defined
  re-usable actions.

Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>